+++ youthmap/src/main/java/com/example/demo/YouthmapApplication.java
@@ -1,11 +1,13 @@
 package com.example.demo;
 
+import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableScheduling
 @SpringBootApplication
+@MapperScan("com.example.demo.dao")
 public class YouthmapApplication {
 
 	public static void main(String[] args) {
